Implement -dry-run switch for read-only operation.

+++ main.go
@@ -19,6 +19,7 @@ import (
 var (
 	apiKey           = flag.String("apikey", os.Getenv("PRODUCTSTATUS_API_KEY"), "Productstatus API key")
 	brokers          = flag.String("brokers", os.Getenv("KAFKA_BROKERS"), "The Kafka brokers to connect to, as a comma separated list")
+	dryRun           = flag.Bool("dry-run", false, "Disable all write operations")
 	offset           = flag.Int64("offset", sarama.OffsetNewest, "Kafka message offset to start reading from")
 	products         = flag.String("products", "", "Which products backends to process expired messages for")
 	productstatusUrl = flag.String("productstatus", os.Getenv("PRODUCTSTATUS_URL"), "URL to the Productstatus web service")
@@ -259,11 +260,16 @@ func rm(path string) error {
 
 // handlePatch updates a DataInstance resource remotely, marking it as deleted.
 func handlePatch(c *productstatus.Client, dataInstance *productstatus.DataInstance) error {
-	err := c.DeleteResource(dataInstance)
-	if err != nil {
-		return fmt.Errorf("Unable to mark resource '%s' as deleted: %s", dataInstance.Resource_uri, err)
+	prefix := ""
+	if *dryRun {
+		prefix = "[DRY-RUN] "
+	} else {
+		err := c.DeleteResource(dataInstance)
+		if err != nil {
+			return fmt.Errorf("Unable to mark resource '%s' as deleted: %s", dataInstance.Resource_uri, err)
+		}
 	}
-	log.Printf("Resource '%s' has been marked as deleted in Productstatus.\n", dataInstance.Resource_uri)
+	log.Printf("%sResource '%s' has been marked as deleted in Productstatus.\n", prefix, dataInstance.Resource_uri)
 	return nil
 }
 
@@ -278,12 +284,17 @@ func handleDelete(dataInstance *productstatus.DataInstance, patch chan *products
 		return fmt.Errorf("%s: productshredder can only delete files with URL scheme 'file'", dataInstance.Url)
 	}
 
-	err = rm(url.Path)
-	if err != nil {
-		return fmt.Errorf("Failed to delete '%s': %s", url.Path, err)
+	prefix := ""
+	if *dryRun {
+		prefix = "[DRY-RUN] "
+	} else {
+		err = rm(url.Path)
+		if err != nil {
+			return fmt.Errorf("Failed to delete '%s': %s", url.Path, err)
+		}
 	}
 
-	log.Printf("Deleted: '%s'\n", url.Path)
+	log.Printf("%sDeleted: '%s'\n", prefix, url.Path)
 
 	patch <- dataInstance
